Updated main.dart and created two new files
Extracted out Scaffold from main, into a new
widget in a new file home_screen_layout.dart

Created a basic multi step form which has the
stepper in vertical direction.

Put the drawer elements as children of ListView
and defined click events on the button, and first
label.

On clicking the Projects the home page changes
to the projects screen (which is also the default
home page). On clicking the New Project button,
the projects screen is replaced by a multi-step
form.