Created a ProjectListing widget
Created of list of projects as list of map
objects
Learned about Table and TableRow widgets, and
used the list of projects to populate elements
in TableRow widgets inside Table widget

Created a tab switching functionality, with the
tabs 'All Projects', 'In progress', 'Completed',
'Hold'. Currently, the switching of the tabs,
will show the name of the tab as the heading
inside the project listing. And, in any tab,
all the projects are displayed.