Added escalation matrix
Created an escalation matrix table,
containing the states of the current selection
of persons.

The dropdown would change the person name,
and accordingly the role is updated, because
the dropdown values are part of the state.

Used three state variables, for different levels,
to ensure we can select different person at each
level.

Currently the same table is displayed, for each
of the matrices.