Changed the approach for navigation to forms
Earlier, the form for new project, was displayed
on the same page, however, now a navigation has
been setup with the proper data and methods
passed into the next route.

This was done for achieving Navigator.pop
on clicking on the Continue button, in the last
form of the multi-step form.

Also, the state has been moved up to HomeScreen,
and listOfProjects is passed to ProjectsScreen,
and updateListOfProjects method is passed to the
CreateProjectScreen. Any actions would update the
state, and it will be visible for all the widgets.